add IAP

+++ app/src/main/java/com/wingsmight/audiorecorder/CloudStoragePlan.java
@@ -1,29 +1,15 @@
 package com.wingsmight.audiorecorder;
 
 public class CloudStoragePlan {
-//    public enum Plan {
-//        case free200MB
-//        case paid2GB
-//
-//
-//        public var size: Int {
-//            switch self {
-//                case .free200MB:
-//                return 200 * 1024 * 1024
-//                case .paid2GB:
-//                return 2 * 1024 * 1024 * 1024
-//            }
-//        }
-//    }
-    public static int getSize(Plan plan) {
+    public static long getSize(Plan plan) {
         switch (plan){
             case free200MB:
-                return 200 * 1024 * 1024;
+                return 200L * 1024L * 1024L;
             case paid2GB:
-                return 2 * 1024 * 1024 * 1024;
+                return 2L * 1024L * 1024L * 1024L;
         }
 
-        return 200 * 1024 * 1024;
+        return 200L * 1024L * 1024L;
     }
 
     public enum Plan {
